backend/requirements.min.txt
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.7.0
python-dotenv==1.0.0
whitenoise
dj-database-url
gunicorn
